plot3.png
Plots
Sub_metering_1: energy sub-metering No. 1 (in watt-hour of active
energy). It corresponds to the kitchen, containing mainly a dishwasher,
an oven and a microwave (hot plates are not electric but gas powered).
Sub_metering_2: energy sub-metering No. 2 (in watt-hour of active
energy). It corresponds to the laundry room, containing a
washing-machine, a tumble-drier, a refrigerator and a light.
Sub_metering_3: energy sub-metering No. 3 (in watt-hour of active
energy). It corresponds to an electric water-heater and an
air-conditioner during a two day interval.